refactor(gui): stay within complexity budgets
